smart-alumni-connect/aadhaar_verification/requirements.txt
numpy==1.26.2
opencv-python==4.8.1.78
pyzbar==0.1.9
cryptography==41.0.7
pillow==10.1.0
Flask==3.0.0